Refactor: Extract shared code from bracket components
Extract duplicated code from Bracket.jsx, SixAllianceBracket.jsx,
FourAllianceBracketFTC.jsx, and FourAllianceBracket.jsx into shared modules:

- bracketConstants.js: Shared color and font weight constants
- bracketHelpers.js: Shared utility functions (getTeamByStation,
  getAllianceNumbersForDisplay, getAllianceNameForDisplay, etc.)
- useBracketState.js: Custom hook for winner selection modal state
- WinnerSelectionModal.jsx: Shared modal component for winner selection

Each bracket component now imports from these shared modules and uses
thin wrapper functions to bind component-specific parameters. All
bracket-specific logic (SVG layout, matchClasses, finals display
helpers, handleConfirmWinner) remains in each component.
