+++ src/components/bracketConstants.js
@@ -1,0 +1,15 @@
+// Shared color and font weight constants for all bracket components
+
+// Ball colors
+export const GOLD = "#FFCA10";
+export const RED = "#FF0000";
+export const BLUE = "#0000FF";
+export const GREEN = "#09BA48";
+export const BLACK = "#000000";
+export const WHITE = "#FFFFFF";
+
+// Font weights
+export const black = "900";
+export const bold = "700";
+export const semibold = "600";
+//export const normal = "400";
